+++ NEWS.md
@@ -3,6 +3,7 @@
 **New features**
 
 - Added `polar` to `canvas_flow()`, thanks to @researchremora on twitter for the idea.
+- Added more color palettes to the `colorPalette()` function.
 
 # aRtsy 0.1.6